bot/pyproject.toml
[project]
name = "injoy-menu-bot"
version = "0.1.0"
description = "Telegram bot for InJoy cafe menu"
requires-python = ">=3.12"
dependencies = [
  "aiogram>=3.20",
  "httpx>=0.28.0",
  "pillow>=11.0.0",
  "pydantic-settings>=2.7.0",
]

[tool.uv]
package = false
